\name{IPMbook-package}
\alias{IPMbook-package}
\alias{IPMbook}
\docType{package}
\encoding{UTF-8}
\title{
Functions and data for the Book 'Integrated Population Modeling'
}
\description{
Provides functions and data sets needed to run the code given in the book \emph{Integrated Population Modeling} by Marc Kéry and Michael Schaub, Academic Press (in preparation).
}
\details{


The functions are listed by chapter below.
}

\section{Chapter X}{
  \describe{
    \item{\code{\link{demoMCMC}}}{ Demonstration of a random-walk sampler for MCMC (section x.x)}
  }
}


%%\section{DATA SETS}{
%%  \describe{
%%%  \item{\code{\link{willowWarbler}}}{ Data for willow warblers in Britain (AHM2-3.4.1)}

%%  } }

\section{UTILITY FUNCTIONS}{
  \describe{
  \item{\code{\link{dbeta2}}}{ Beta distribution with mean and sd.}
  \item{\code{\link{dbeta3}}}{ Beta distribution with mode and concentration.}
  \item{\code{\link{dgamma2}}}{ Gamma distribution with mean and sd.}
   } }


\author{
Michael Schaub, Marc Kéry, Mike Meredith
}

\keyword{package}
